:white_check_mark: cover integration of delegate details page
